+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 <Contributors, please add your changes below this line>
 
-* Adds personalization `get` and `set` strategy method - PR [#395](https://github.com/algolia/algoliasearch-client-python/pull/395)
+* Adds `get` and `set` personalization strategy methods - PR [#395](https://github.com/algolia/algoliasearch-client-python/pull/395)
 
 ### 1.18.1 - 2018-11-26